+++ Sources/MapboxNavigation/StyleManager.swift
@@ -13,11 +13,12 @@ public protocol StyleManagerDelegate: AnyObject, UnimplementedLogging {
     func location(for styleManager: StyleManager) -> CLLocation?
     
     /**
-     Asks the delegate for the view to be used when refreshing appearance. 
+     Asks the delegate for the view to be used when refreshing appearance.
      
      The default implementation of this method will attempt to cast the delegate to type
      `UIViewController` and use its `view` property.
      */
+    @available(*, deprecated, message: "All views appearance will be refreshed without the `UITextEffectsWindow`.")
     func styleManager(_ styleManager: StyleManager, viewForApplying currentStyle: Style?) -> UIView?
     
     /**
@@ -56,10 +57,11 @@ public extension StyleManagerDelegate {
         logUnimplemented(protocolType: StyleManagerDelegate.self, level: .debug)
     }
     
+    @available(*, deprecated, message: "All views appearance will be refreshed without the `UITextEffectsWindow`.")
     func styleManager(_ styleManager: StyleManager, viewForApplying currentStyle: Style?) -> UIView? {
-        // Short-circuit refresh logic if the view hasn't yet loaded since we don't want the `self.view` 
+        // Short-circuit refresh logic if the view hasn't yet loaded since we don't want the `self.view`
         // call to trigger `loadView`.
-        if let vc = self as? UIViewController, vc.isViewLoaded { 
+        if let vc = self as? UIViewController, vc.isViewLoaded {
             return vc.view
         }
         
@@ -177,9 +179,9 @@ open class StyleManager {
      Applies the `Style` with type matching `type`and notifies `StyleManager.delegate` upon completion. 
      */
     public func applyStyle(type styleType: StyleType) {
-        guard currentStyleType != styleType else { return }
-        
-        NSObject.cancelPreviousPerformRequests(withTarget: self, selector: #selector(timeOfDayChanged), object: nil)
+        if currentStyleType != styleType {
+            NSObject.cancelPreviousPerformRequests(withTarget: self, selector: #selector(timeOfDayChanged), object: nil)
+        }
         
         for style in styles {
             if style.styleType == styleType {
@@ -257,15 +259,16 @@ open class StyleManager {
         forceRefreshAppearance()
     }
     
-    // workaround to refresh appearance by removing the view and then adding it again
+    // Workaround to refresh appearance by removing all views and then adding them again.
+    // UITextEffectsWindow will be created when system keyboard is shown and cannot be safely removed.
     func forceRefreshAppearance() {
-        if 
-            let view = delegate?.styleManager(self, viewForApplying: currentStyle), 
-            let superview = view.superview, 
-            let index = superview.subviews.firstIndex(of: view) 
-        {
-            view.removeFromSuperview()
-            superview.insertSubview(view, at: index)
+        for window in UIApplication.shared.windows {
+            if !window.isKind(of: NSClassFromString("UITextEffectsWindow") ?? NSString.classForCoder()) {
+                for view in window.subviews {
+                    view.removeFromSuperview()
+                    window.addSubview(view)
+                }
+            }
         }
         
         delegate?.styleManagerDidRefreshAppearance(self)